+++ test/util.c
@@ -15,7 +15,7 @@
 
 uint8_t *
 mk_stub(struct Sqsh *sqsh, uint8_t *payload, size_t payload_size,
-		uint64_t *target_size) {
+		size_t *target_size) {
 	struct SqshDataSuperblock superblock = {
 			.magic = SQSH_SUPERBLOCK_MAGIC,
 			.inode_count = 0,
